requirements.txt
fastapi
uvicorn
mediapipe==0.10.11
opencv-python-headless
python-multipart
pillow